Vehiculo y Usuario
Faltan solucionar fallas

+++ TDA.Entities/Usuarios.cs
@@ -24,7 +24,7 @@ namespace TDA.Entities
         [DataMember]
         public string RolNombre { get; set; }
         [DataMember]
-        public int? Empleado { get; set; }
+        public long? Empleado { get; set; }
         [DataMember]
         public long? UsuarioAlta { get; set; }
         [DataMember]
